+++ bind.js
@@ -1,0 +1,10 @@
+module.exports = function (obj, prop) {
+  return {
+    get: function () {
+      return obj[prop];
+    },
+    set: function (value) {
+      obj[prop] = value;
+    }
+  };
+};
